requirements.txt
torch>=2.2.0
pytorch-lightning>=2.2.0
transformers>=4.37.0
datasets>=2.15.0
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.3.0
pyyaml>=6.0.1
hydra-core>=1.3.2
requests>=2.31.0
tqdm>=4.66.0
biopython>=1.83
mdtraj>=1.9.9
pyarrow>=14.0.0
fastapi>=0.108.0
uvicorn[standard]>=0.25.0
rich>=13.7.0
mlflow>=2.9.0
weaviate-client>=4.4.0
wandb>=0.16.0
python-dotenv>=1.0.0
polars>=0.20.0
